refactor(api): move controller method
getCurrentUser
