+++ api/tests/identity-access-management/unit/application/user.controller.test.js
@@ -14,6 +14,27 @@ describe('Unit | Identity Access Management | Application | Controller | User', 
     };
   });
 
+  describe('#getCurrentUser', function () {
+    it('gets the current user', async function () {
+      // given
+      const request = { auth: { credentials: { userId: 1 } } };
+      const currentUser = Symbol('current-user');
+      const getCurrentUserStub = sinon.stub(usecases, 'getCurrentUser');
+      const userWithActivitySerializer = { serialize: sinon.stub() };
+
+      usecases.getCurrentUser.withArgs({ authenticatedUserId: 1 }).resolves(currentUser);
+      userWithActivitySerializer.serialize.withArgs(currentUser).returns('ok');
+
+      // when
+      const response = await userController.getCurrentUser(request, hFake, { userWithActivitySerializer });
+
+      // then
+      expect(response).to.be.equal('ok');
+      expect(getCurrentUserStub).to.have.been.calledWithExactly({ authenticatedUserId: 1 });
+      expect(userWithActivitySerializer.serialize).to.have.been.calledWithExactly(currentUser);
+    });
+  });
+
   describe('#save', function () {
     const email = 'to-be-free@ozone.airplane';
     const password = 'Password123';
